`BatchQuery` refactored
