+++ src/vector/index.ts
@@ -1,2 +1,1 @@
 export * from './query';
-export * from './utils';
